added two new folders and moved files and  added a Card.js component as a custom component
