--- src/components/Expenses.css
@@ -1,9 +1,0 @@
-.expenses {
-    padding: 1rem;
-    background-color: rgb(31, 31, 31);
-    margin: 2rem auto;
-    width: 50rem;
-    max-width: 95%;
-    border-radius: 12px;
-    box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
-  }
