+++ src/components/Expenses/ExpenseItem.js
@@ -1,19 +1,20 @@
 import ExpenseDate from "./ExpenseDate";
 import "./ExpenseItem.css";
+import Card from '../UI/Card'
 
-function ExpenseItem(props) {
+const ExpenseItem = (props) => {
   // parameter for passing "props"
 
   //  console.log("Date", props.date);
 
   return (
-    <div className="expense-item">
+    <Card className="expense-item">
       <ExpenseDate date={props.date} />
       <div className="expense-item__description">
         <h2>{props.title}</h2>
         <div className="expense-item__price">${props.amount}</div>
       </div>
-    </div>
+    </Card>
   );
 }
