+++ src/components/UI/Card.js
@@ -1,0 +1,8 @@
+import "./Card.css";
+
+const Card = (props) => {
+    const classes = 'card ' + props.className;
+  return <div className={classes}>{props.children}</div>;
+}
+
+export default Card;
